Comments and readme
some final touches before submitting

+++ vagrant/catalog/application.py
@@ -1,7 +1,7 @@
 from flask import Flask, render_template, request, redirect, jsonify, url_for, flash
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
-from dbsetup import Base, Category, Item
+from dbsetup import Base, Category, Item, User
 #imports for OAuth
 from flask import session as login_session
 import random
@@ -13,16 +13,12 @@ import json
 from flask import make_response
 import requests
 
-#Not sure what this bit does.
 app = Flask(__name__)
-
 
 CLIENT_ID = json.loads(open('client_secrets.json', 'r').read())['web']['client_id']
 APPLICATION_NAME = "Catalog Application"
 
-
-
-
+#Get the db interface going
 engine = create_engine('sqlite:///catalog.db')
 Base.metadata.bind = engine
 
@@ -36,18 +32,19 @@ def showLogin():
     state = ''.join(random.choice(string.ascii_uppercase + string.digits)
                     for x in xrange(32))
     login_session['state'] = state
-    # return "The current session state is %s" % login_session['state']
+    #next commented line is some debug code 
+	#return "The current session state is %s" % login_session['state']
     return render_template('login.html', STATE=state)
 
-
+# Connect a user using google's OAuth
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
-    # Validate state token
+    # Validate the state token
     if request.args.get('state') != login_session['state']:
         response = make_response(json.dumps('Invalid state parameter.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
-    # Obtain authorization code
+    # Obtain an authorization code
     code = request.data
 
     try:
@@ -61,13 +58,13 @@ def gconnect():
         response.headers['Content-Type'] = 'application/json'
         return response
 
-    # Check that the access token is valid.
+    # Check that the access token is actually valid.
     access_token = credentials.access_token
     url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
            % access_token)
     h = httplib2.Http()
     result = json.loads(h.request(url, 'GET')[1])
-    # If there was an error in the access token info, abort.
+    # If there was an error in the access token info, jump ship.
     if result.get('error') is not None:
         response = make_response(json.dumps(result.get('error')), 500)
         response.headers['Content-Type'] = 'application/json'
@@ -100,7 +97,7 @@ def gconnect():
     login_session['credentials'] = credentials
     login_session['gplus_id'] = gplus_id
 
-    # Get user info
+    # Get user deets
     userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
     params = {'access_token': credentials.access_token, 'alt': 'json'}
     answer = requests.get(userinfo_url, params=params)
@@ -110,6 +107,13 @@ def gconnect():
     login_session['username'] = data['name']
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
+    
+    #check if user exists before creating a new one
+    user_id = getUserId(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
 
     output = ''
     output += '<h1>Welcome, '
@@ -122,6 +126,7 @@ def gconnect():
     print "done!"
     return output
 
+#Disconnect a connected google OAuth user
 @app.route('/gdisconnect')
 def gdisconnect():
 	credentials = login_session.get('credentials')
@@ -141,21 +146,44 @@ def gdisconnect():
 	print result
 	print 'status is'
 	print result['status']
+	#if the logout worked, set a flash message and delete all of the stored information in login_session
 	if result['status'] == '200':
+		flash("you are now logged out %s" % login_session['username'])
 		del login_session['credentials']
 		del login_session['gplus_id']
 		del login_session['username']
 		del login_session['email']
-		del login_session['picture']
-		response = make_response(json.dumps('Successfully disconnected.'), 200)
-		response.headers['Content-Type'] = 'application/json'
-		return response
+		del login_session['picture']		
+		return redirect ('/catalog')
 	else:
+	#if the logout failed, let the user know. This shouldn't really occur in the wild much.  
+	#Should stress test this in reality as not being able to log out would be extraordinarily frustrating (
+	#e.g. using a public library computer and running out of time)
 		response = make_response(json.dumps('Failed to revoke token for given user.'), 400)
 		response.headers['Content-Type'] = 'application/json'
 		return response
-	
-	
+
+#Grab a user's ID
+def getUserId(email):
+	try:
+		user = session.query(User).filter_by(email = email).one()
+		return user.id
+	except:
+		return None
+
+#Grab a user's stored information
+def getUserInfo(user_id):
+	user = session.query(User).filter_by(id = user_id).one()
+	return user
+		
+#Create a new user in a general way that will support multiple OAuth providers
+def createUser(login_session):
+	newUser = User(name = login_session['username'], email = login_session['email'], picture = login_session['picture'])
+	session.add(newUser)
+	session.commit()
+	user = session.query(User).filter_by(email = login_session['email']).one()
+	return user.id
+
 #Making an API Endpoint (GET request)
 @app.route('/category/<path:category_name>/items/JSON/')
 def categoryJSON(category_name):
@@ -170,31 +198,49 @@ def itemJSON(category_name, item_id, item_name):
 	
 #main routing
 @app.route('/')
+@app.route('/index')
 @app.route('/category/')
 @app.route('/catalog/')
 def index():
 	categories = session.query(Category).all()
 	allitems = session.query(Item).join(Item.category).order_by(Item.id).all()
-	return render_template('index.html', categories = categories, allitems = allitems)
+	if 'username' not in login_session:
+		return render_template('publicindex.html', categories = categories, allitems = allitems)
+	else:
+		return render_template('index.html', categories = categories, allitems = allitems)
 	
+	
+#Routing for categories. Using unique category names otherwise this routing would break
 @app.route('/category/<path:category_name>/items/')
 def category(category_name):
 	category = session.query(Category).filter_by(name = category_name).one()
 	items = session.query(Item).filter_by(category_id = category.id).all()
-	return render_template('category.html', category=category, items=items)
+	creator = getUserInfo(category.user_id)
+	if 'username' not in login_session or creator.id != login_session['user_id']:
+		return render_template('publiccategory.html', category=category, items=items)
+	else:
+		return render_template('category.html', category=category, items=items, creator = creator)
 	
+	
+#Routing for items.  Since multiple items of same name could exist, am using item_id in the route to prevent bugs since that's unique.
 @app.route('/category/<path:category_name>/<int:item_id>/<path:item_name>/')
 def item(category_name, item_id, item_name):
 	item = session.query(Item).filter_by(id = item_id).one()
 	category = session.query(Category).filter_by(name = category_name).one()
-	return render_template('item.html', item=item, category=category)
+	creator = getUserInfo(category.user_id)
+	if 'username' not in login_session or creator.id != login_session['user_id']:
+		return render_template('publicitem.html', item=item, category=category)
+	else:
+		return render_template('item.html', item=item, category=category)
 	
+	
+#making a new Category
 @app.route('/category/new/', methods=['GET','POST'])
 def newCategory():
 	if 'username' not in login_session:
 		return redirect('/login')
 	if request.method=='POST':
-		newCategory=Category(name= request.form['name'])
+		newCategory=Category(name= request.form['name'], user_id = login_session['user_id'])
 		session.add(newCategory)
 		session.commit()
 		flash("The category "+newCategory.name+" has been created!")
@@ -202,14 +248,15 @@ def newCategory():
 	else:
 		return render_template('newcat.html')
 
-#@app.route('/category/edit/', methods=['GET','POST'])
-#def editCat(category_name):
+#editing an existing category
 @app.route('/category/<path:category_name>/edit/', methods=['GET','POST'])
 def editCategory(category_name):
 	if 'username' not in login_session:
 		return redirect('/login')
 	categorytoedit = session.query(Category).filter_by(name= category_name).one()
 	oldname=categorytoedit.name
+	if categorytoedit.user_id != login_session['user_id']:
+		return "<script>function myFunction() {alert('You are not authorized to edit this category. Please create your own category in order to edit.')}</script><body onload = 'myFunction()''>"
 	if request.method=='POST':
 		if request.form['name']:
 			categorytoedit.name = request.form['name']
@@ -220,12 +267,14 @@ def editCategory(category_name):
 	else:
 		return render_template('editcat.html', category=categorytoedit)
 
+#Deleting a category
 @app.route('/category/<path:category_name>/delete/', methods=['GET','POST'])
 def deleteCat(category_name):
+	categorytodelete = session.query(Category).filter_by(name= category_name).one()
 	if 'username' not in login_session:
 		return redirect('/login')
-	categorytodelete = session.query(Category).filter_by(name= category_name).one()
-	
+	if categorytodelete.user_id != login_session['user_id']:
+		return "<script>function myFunction() {alert('You are not authorized to delete this category. Please create your own category in order to delete.')}</script><body onload = 'myFunction()''>"
 	if request.method=='POST':
 		session.delete(categorytodelete)
 		session.commit()
@@ -234,13 +283,14 @@ def deleteCat(category_name):
 	else:
 		return render_template('deletecat.html', category=categorytodelete)
 		
+#making a new item
 @app.route('/category/<path:category_name>/new/', methods=['GET','POST'])
 def newItem(category_name):
 	if 'username' not in login_session:
 		return redirect('/login')
 	if request.method=='POST':
 		category = session.query(Category).filter_by(name = category_name).one()
-		newItem=Item(name= request.form['name'], price= (request.form['price']), description= request.form['description'], category_id = category.id)
+		newItem=Item(name= request.form['name'], price= (request.form['price']), description= request.form['description'], category_id = category.id, user_id = login_session['user_id'])
 		session.add(newItem)
 		session.commit()
 		flash(newItem.name+" has been created for the "+category.name+" category!")
@@ -249,6 +299,7 @@ def newItem(category_name):
 		category = session.query(Category).filter_by(name = category_name).one()
 		return render_template('newitem.html', category_name=category_name, category=category)
 	
+#editing an existing item
 @app.route('/category/<path:category_name>/<int:item_id>/edit/', methods=['GET','POST'])
 def editItem(category_name, item_id):
 	if 'username' not in login_session:
@@ -256,6 +307,8 @@ def editItem(category_name, item_id):
 	category = session.query(Category).filter_by(name= category_name).one()
 	itemtoedit = session.query(Item).filter_by(id = item_id).one()
 	oldname = itemtoedit.name
+	if itemtoedit.user_id != login_session['user_id']:
+		return "<script>function myFunction() {alert('You are not authorized to edit this item. Please create your own items in order to edit.')}</script><body onload = 'myFunction()''>"
 	if request.method=='POST':
 		if request.form['name']:
 			itemtoedit.name = request.form['name']
@@ -270,12 +323,15 @@ def editItem(category_name, item_id):
 	else:
 		return render_template('edit.html', category=category, item=itemtoedit)
 	
+#deleting an existing item
 @app.route('/category/<path:category_name>/<int:item_id>/delete/', methods=['GET','POST'])
 def deleteItem(category_name, item_id):
 	if 'username' not in login_session:
 		return redirect('/login')
 	category = session.query(Category).filter_by(name= category_name).one()
 	itemtodelete = session.query(Item).filter_by(id= item_id).one()
+	if itemtodelete.user_id != login_session['user_id']:
+		return "<script>function myFunction() {alert('You are not authorized to delete this item. Please create your own items in order to delete.')}</script><body onload = 'myFunction()''>"
 	if request.method=='POST':
 		session.delete(itemtodelete)
 		session.commit()
